PDF Service Points Fix
Service points were only being shown if there were employees also being shown. Adjusted the next if statement so that it checked if service_points.count was less than 1 it could skip the section. Added if statements around the each to insure we didn't try to render service points or employees that had a results.